Jump Mechanics
Added jump mechanics and moved the camera a bit, had to do a couple of things differently from how it would work in 2d. Not sure how good it feels, but it works for now. Don't exactly know how to make it look like the player is charging the jump, probably would be through animation. :D