Normal Walls, Platform Prefab, Collider Material, Camera Offset, and Jump Height Changes
We could do invisible walls, I had them in the code but remembered that you bounce off of walls and platform sides in jump king. So, restricting the player through transform wouldn't work. Added visible walls, we could change them to be invisible but I think it's good to let the player know where the bounds are in some way. Made a very basic platform prefab, eventually we can make a slope prefab as well. Added a material to the collider that prevents the player from sticking to the sides of walls/platforms by setting the friction to 0 and combine to minimum. For this material, might need to mess around with bounciness to bounce off of walls, or do that in code so it's only while the player is in the air. Added a camera offset so you can see the platforms above you more clearly. Messed with jump height to make it feel a little less floaty.