Fan Prefab and WindArea script
Added a WindArea script that could be used on any collider that we want to have wind, just have to add the script, change its tag, then set the strength and direction. I also added a fan prefab for a vertical fan. The particles can be changed just thought I'd put some temporary ones on there for visuals. The strength/direction can also be tweaked depending on where we want to put the fan or change its orientation.  :D