Menu/Save/In progress pause
Added a menu where you can start in different areas. Added save to make that possible. Am in the process of adding a pause screen that does the same sort of thing. If eventually we need to load the art scene I added a function for that as well, its commented out for now as it spawns an extra player if loaded.